requirements.txt
# Core dependencies
python-telegram-bot>=20.6
asyncio

# Database
supabase>=2.0.0

# Computer Vision & Image Processing
opencv-python>=4.8.1
Pillow>=10.0.0
numpy>=1.24.3

# OCR for document text extraction
pytesseract>=0.3.10

# HTTP requests
requests>=2.31.0

# Text processing
regex>=2023.8.8

# Logging
colorama>=0.4.6

# Data validation and type hints
typing-extensions>=4.7.1

# Image processing utilities
scikit-image>=0.21.0

# Audio processing (optional - for future audio deepfake detection)
librosa>=0.10.1
soundfile>=0.12.1

# Video processing utilities
imageio>=2.31.3

# Machine Learning (optional - for enhanced deepfake detection)
scikit-learn>=1.3.0

# Date/time utilities
python-dateutil>=2.8.2

# File handling
pathlib

# JSON processing
jsonschema>=4.19.0

# Environment configuration
python-dotenv>=1.0.0

# Performance monitoring
psutil>=5.9.5

# Hash utilities (built-in)
hashlib

# Statistical computations
scipy>=1.11.2

# PDF processing (for document analysis)
PyPDF2>=3.0.1
pdfminer.six>=20221105

# Excel file processing (optional)
openpyxl>=3.1.2

# Word document processing (optional)
python-docx>=0.8.11

# Image hashing for duplicate detection
imagehash>=4.3.1

# Natural Language Processing (optional - for advanced text analysis)
nltk>=3.8.1

# Progress bars for long operations
tqdm>=4.66.1

# Configuration file parsing
configparser

# Cryptographic functions
cryptography>=41.0.4

# System information
platform

# Memory optimization
gc